Day Calendar: Refactor to tsx component
