+++ package.json
@@ -76,7 +76,7 @@
     "react-redux": "^7.2.6",
     "react-render-html": "^0.1.6",
     "react-select": "^1.3.0",
-    "react-swipeable": "^4.3.2",
+    "react-swipeable": "^5.5.1",
     "react-test-renderer": "^16.14.0",
     "react-tooltip": "^4.2.21",
     "reactify": "^1.1.1",
@@ -137,6 +137,7 @@
     "@types/clipboard": "^2.0.7",
     "@types/jest": "^27.0.3",
     "@types/js-cookie": "^3.0.1",
+    "@types/react-swipeable": "^5.2.0",
     "@types/react-tooltip": "^4.2.4",
     "@types/tinycolor2": "^1.4.3",
     "eslint-config-prettier": "^8.3.0",
